+++ src/api/projects/variables/create.rs
@@ -62,6 +62,9 @@ pub struct CreateProjectVariable<'a> {
     /// Whether the variable is masked.
     #[builder(default)]
     masked: Option<bool>,
+    /// Whether the variable is treated as a raw string.
+    #[builder(default)]
+    raw: Option<bool>,
     /// The environment scope of the variable.
     #[builder(setter(into), default)]
     environment_scope: Option<Cow<'a, str>>,
@@ -92,6 +95,7 @@ impl<'a> Endpoint for CreateProjectVariable<'a> {
             .push_opt("variable_type", self.variable_type)
             .push_opt("protected", self.protected)
             .push_opt("masked", self.masked)
+            .push_opt("raw", self.raw)
             .push_opt("environment_scope", self.environment_scope.as_ref());
 
         params.into_body()
@@ -252,6 +256,27 @@ mod tests {
             .key("testkey")
             .value("testvalue")
             .masked(true)
+            .build()
+            .unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_raw() {
+        let endpoint = ExpectedUrl::builder()
+            .method(Method::POST)
+            .endpoint("projects/simple%2Fproject/variables")
+            .content_type("application/x-www-form-urlencoded")
+            .body_str(concat!("key=testkey", "&value=testvalue", "&raw=true"))
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = CreateProjectVariable::builder()
+            .project("simple/project")
+            .key("testkey")
+            .value("testvalue")
+            .raw(true)
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();